requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
eventlet==0.36.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.35
alembic==1.13.2
python-dotenv==1.0.1
passlib==1.7.4
email-validator==2.2.0
